+++ Utility/UTFUtils.cpp
@@ -8,7 +8,6 @@
 #include "StringUtils.h"
 #include "StringFormat.h"
 #include <stdlib.h>
-#include <malloc.h>
 
 namespace Utility
 {
@@ -663,28 +662,28 @@ int utf8_read_escape_sequence(char* str, uint32* dst)
         do {
             digs[dno++] = str[i++];
         } while (octal_digit(str[i]) && dno < 3);
-        ch = strtol(digs, NULL, 8);
+        ch = (uint32)strtol(digs, NULL, 8);
     }
     else if (str[0] == 'x') {
         while (XlIsHex(str[i]) && dno < 2) {
             digs[dno++] = str[i++];
         }
         if (dno > 0)
-            ch = strtol(digs, NULL, 16);
+            ch = (uint32)strtol(digs, NULL, 16);
     }
     else if (str[0] == 'u') {
         while (XlIsHex(str[i]) && dno < 4) {
             digs[dno++] = str[i++];
         }
         if (dno > 0)
-            ch = strtol(digs, NULL, 16);
+            ch = (uint32)strtol(digs, NULL, 16);
     }
     else if (str[0] == 'U') {
         while (XlIsHex(str[i]) && dno < 8) {
             digs[dno++] = str[i++];
         }
         if (dno > 0)
-            ch = strtol(digs, NULL, 16);
+            ch = (uint32)strtol(digs, NULL, 16);
     }
     *dst = ch;
